changed F matrix to MatrixXi in .h/.cpp files, commented out cross products

+++ include/anisotropic_measure.h
@@ -18,7 +18,7 @@
 */
 
 void anisotropic_measure(
-  const Eigen::MatrixXd & F,
+  const Eigen::MatrixXi & F,
   const Eigen::MatrixXd & V,
   const Eigen::MatrixXd & N,
   const Eigen::MatrixXd & A_ratio_at_v,
